requirements.txt
aniso8601==8.0.0
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
protobuf==3.12.2
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
pyrsistent==0.16.0
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
